update: 更新文件 setvoicemailinfo_fuzzer.cpp

+++ test/fuzztest/setvoicemailinfo_fuzzer/setvoicemailinfo_fuzzer.cpp
@@ -111,10 +111,10 @@ void SetVoiceMailInfo(const uint8_t *data, size_t size)
     }
 
     int32_t slotId = static_cast<int32_t>(*data % SLOT_NUM);
-    int32_t offset = 0;
+    size_t offset = 0;
     std::string mailNumber(reinterpret_cast<const char *>(data), size);
     offset += sizeof(int32_t);
-    offset = (offset > static_cast<int32_t>size) ? static_cast<int32_t>size : offset;
+    offset = (offset > size) ? size : offset;
     std::string mailName(reinterpret_cast<const char *>(data + offset), size - offset);
     auto mailNameU16 = Str8ToStr16(mailName);
     auto mailNumberU16 = Str8ToStr16(mailNumber);
